feat($AuthCenter): support admin authorization

+++ common/src/main/java/com/jmsoftware/maf/common/domain/authcenter/permission/GetPermissionListByRoleIdListResponse.java
@@ -29,7 +29,7 @@ public class GetPermissionListByRoleIdListResponse {
         /**
          * The Type.
          */
-        private Integer type;
+        private Byte type;
         /**
          * The Permission expression.
          */
